requirements.txt
fastapi
uvicorn
playwright
google-generativeai
openai
python-dotenv
jinja2
python-multipart
protobuf